feat: initialize frontend with Vite and React setup
- Added package.json with dependencies for React, React Router, Axios, and Vite.
- Created initial Vite configuration file.
- Implemented main application structure with routing for Login, SignUp, and Dashboard components.
- Added authentication context for managing user state and token.
- Developed Login and SignUp components for user authentication.
- Created ProtectedRoute component to guard access to the Dashboard.
- Styled application with CSS for a better user experience.
- Included SVG logos for React and Vite in the public assets.

+++ backend/src/repositories/transaction.repo.js
@@ -7,13 +7,16 @@ export const create = async (userid, amount, categoryid, is_partial, datetime) =
     return rows;
 }
 
-export const fetchTransactions = async (userid) => {
+export const fetchTransactions = async (userid, is_partial = null) => {
     const pool = connectDB();
-    const [rows] = await pool.query(`SELECT * FROM transactions WHERE userid = ?`, [userid]);
+    const values = [userid];
+    const [rows] = await pool.query(`
+        SELECT * FROM transactions WHERE userid = ? ${is_partial ? "and is_partial = ?" : ""}`, 
+        values);
     return rows;
 }
 
-export const updateRow = async (transactionid, amount, categoryid, is_partial, date) => {
+export const updateRow = async (userid, transactionid, amount, categoryid, is_partial, date) => {
     const pool = connectDB();
     const fields = [], values = [];
 
@@ -25,17 +28,18 @@ export const updateRow = async (transactionid, amount, categoryid, is_partial, d
     const sql = `
     UPDATE transactions SET 
     ${fields.join(", ")} 
-    WHERE transactionid = ?
+    WHERE transactionid = ? 
+    AND userid = ?
     `;
 
-    values.push(transactionid);
+    values.push(transactionid, userid);
 
     const [rows] = await pool.query(sql, values);
     return rows;
 }
 
-export const deleteRow = async (transactionid) => {
+export const deleteRow = async (userid, transactionid) => {
     const pool = connectDB();
-    const [rows] = await pool.query(`DELETE FROM transactions WHERE transactionid = ?`, [transactionid]);
+    const [rows] = await pool.query(`DELETE FROM transactions WHERE transactionid = ? AND userid = ?`, [transactionid, userid]);
     return rows;
 }
